backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.7
matplotlib>=3.7.0
numpy>=1.24.0